+++ Tema 3/Ejercicio03.java
@@ -8,7 +8,7 @@
    public static void main (String[] arg) {
    
     System.out.print("Por favor, introduzca la cantidad de pesetas que quiere convertir: ");
-    double pesetas = Integer.parseInt(System.console().readLine());
+    int pesetas = Integer.parseInt(System.console().readLine());
 
     double euros = pesetas / 166.386;
